Write commit messages with emacs client
Instead of spinning up a new instance of emacs, just use
emacsclient. Note that this will complain if the server side isn't
running, but that can be overcome by running `M-x server-start`. This
can probably be done in `init.el` but I haven't bothered to figure that
out just yet.